+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - add: modality support
 - add: folder based on db table prefix
+- add: support for parquet files and Delta/Hive/Iceberg directories
 
 ## 0.2.1 (2026-01-21)